+++ core/src/test/kotlin/io/dwsoft/restx/RestXConfigurationTests.kt
@@ -1,14 +1,15 @@
 package io.dwsoft.restx
 
-import io.dwsoft.restx.fault.cause.code
-import io.dwsoft.restx.fault.cause.message
-import io.dwsoft.restx.fault.payload.MultiErrorPayload
-import io.dwsoft.restx.fault.payload.OperationError
-import io.dwsoft.restx.fault.payload.RequestDataError
-import io.dwsoft.restx.fault.payload.Source
-import io.dwsoft.restx.fault.response.HttpStatus
-import io.dwsoft.restx.fault.response.ResponseGenerator
-import io.dwsoft.restx.fault.response.status
+import io.dwsoft.restx.core.dummy
+import io.dwsoft.restx.core.cause.code
+import io.dwsoft.restx.core.cause.message
+import io.dwsoft.restx.core.payload.MultiErrorPayload
+import io.dwsoft.restx.core.payload.OperationError
+import io.dwsoft.restx.core.payload.RequestDataError
+import io.dwsoft.restx.core.payload.Source
+import io.dwsoft.restx.core.response.HttpStatus
+import io.dwsoft.restx.core.response.ResponseGenerator
+import io.dwsoft.restx.core.response.status
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldContainInOrder
 import io.kotest.matchers.shouldBe
